Address Validation
Block/Suggestions
Update shouldShowSuggestions method to check for suggestion "freshness". This keeps suggestions from being shown more than once, preventing suggestions from a previous address form being shown on other address pages. As the "freshness" will change during the request as suggestions are gathered, the results are cached on the instance to keep a consisten value througout the lifetime of the instance.
Add a method to get a JSON representation of address data that may be changed by EB2C address validation.

Helper/Data
Make sure the region_id pulled from the response message is always an int. This is necessary for the address object to properly look up the region name/data for a given region id.

Model/Suggestion/Group
Flush out this model a bit more. No longer simply a typed varien object.
Model tracks the "freshness" of suggestion data. By default, whenever suggestion data is requested, the data is no longer considered "fresh". This gets used by the Suggestions block to ensure it is only showing fresh suggestions to the user and not those from a previous address validation attempt.
Update and fix magic method doc comments.

Model/Validator
More thorough checks on whether or not an address needs to be validated.
Checks for:
- Addresses previously validated
- Addresses for virtual orders
- Addresses used as a billing address only
- Addresses being saved in the user's address book
Always and more accurately update the validation data in the customer session after validating an address.
Remove the clone/merge functionality. This has been replaced in 2 parts.
1) Only copy over the necessary fields from the original address object to the one create from the address validation response.
2) Instead of keeping a full cloned/merged address in the session, keep a simplified address object with only the data that was actually validated. Simplified address object used to test for a previously validated address.
Make use of the session's "freshness" flag. Add a means of getting the flag from the session object. Make sure the flag doesn't get reset when simply checking for suggestions to exist. Set the flag appropriately when adding new suggestions to the session.

Tests:
Update some existing tests for new changes. Many more tests are required, especially for the validator model.

Frontend - template/JS:
Add a JSON representation of the address data for each suggestion to the input element for that suggestions. Add JS to use this to update the address fields when making suggestions.
Make sure the checkout onSave callbacks/nextStep methods are always updated.