EBC-130 - Don't make allocation request when no items need to be allocated
Consolidate checks to determine if an allocation request is required. Add an additional check to this for ensuring that at least one item in the quote to be allocated has managed stock.
Bit of cleanup to go with it:
Inventory abstract model can inherit from Varien_Object instead of Mage_Core_Model_Abstract.
No need to instantiate a new inventory allocation model for each call to the model in the observer methods. Instead, assign one allocation model and reuse it.
Change some of the getMock calls to getModelMock calls.