EBC-122 - Ensure the suggestions block is added to OPC when an address is invalid but does not have any suggestions.
When an address is invalid but has no suggestions, the suggestions should show up during OPC allowing the user to select the address entered as being correct or enter a new address. This was failing previously due to the observer adding the suggestions in OPC only when there were suggestions other than original or new.
Update some tests to reflect checking for the address to be invalid before adding suggestions instead of having multiple suggestions. Some of these tests can probably be cleaned up as there's a good deal of duplication between some of them.
Some minor clean up work in address:
Inherit Varien_Object instead of Mage_Core_Model_Abstract.
Don't assign helpers as object properties, there's no need for that.
Clean up some log messages.
Add coverage for the response model. There is now one test that is primarilly concerned with the correct validity being reported based on the status code, which passes and covers all statuses, as well as one that will need to ensure messages are logged, which is still marked incomplete.