EBC-141 - item filtering, details & allocation request building, quantity itemId -> sku not item id
Method for filtering managed stock items changed to accept array of quote items instead of the quote, allowing for more flexibility.
Remove unnecessary try/catch in request building methods for details and allocation. Nothing within the block should be expected to thrown an exception.
Replace some loops over quote items with calls to Quote::getItemById
Ensure the quote has a shipping address before trying to build out details and allocation request item nodes. Shipping information is required in the request so not possible to build a valid QuoteItem node without it.
In the quantity response handling, the itemId in the response was being matched against the item id instead of the item sku, which is not correct. Fixed this.
Replaced some not-so-helpful tests with, hopefully, more helpful ones. The tests are a bit of a mess in terms of keeping them DRY and using useful data providers and fixtures but at least the assertsions are meaningful.