EBC-245 - fix up payment method suppression and auto-enabling/disabling of payment methods when enabling/disabling eb2c payments
Replace config lookup via DB to looking for config in the loaded config model, allows for getting config that isn't stored in the DB.
Two lists of payment methods, those exclusive to eb2c which will be enabled/disabled when enabling/disabling eb2c payments, and whitelisted payments which are allowed with eb2c payments and but will not be auto-enabled/disabled (sort of).
Auto-enable free payments when enabling eb2c payments (the sort of mentioned previously).
Add modman links for the adminhtml layout and templates.