EBC-293 Remove env and region config, add hostname
Add a "hostname" config to the admin for the api hostname.
Remove config for api env and region.
Update a bunch of fixtures and tests to match new hostname building
mechanism. Ideally, most of the impacted tests shouldn't have been
failing due to this but that cleanup will come in time.