EBC-340 Move tax helper to not-an-override
Move methods that were on the override helper to the eb2ctax helper.
Delete the override class.
Update references to use the proper helper, 'tax' or 'eb2ctax'.
Some tests were mocking the config registry but a non-mocked one was being stored and used on the helper being used. Instead, the helper methods were mocked which not only got around the issue but is also a better way of doing this - mock the interface, not the implementation.