EBC-363 Check Default Translation Before Adding
Added tests for a function that checks if the default translation exists.
Added a function to check if the default translation exists.
Added a test for _applyDefaultTranslations
Updated _applyDefaultTranslations to use _hasDefaultTranslation
Added a test for the case where parseTranslations is given a null
Altered parseTranslations to accept a null value