US6994 Magento-as-PIM
Generate a MageMaster feed that will contain all product data required for
export. This feed should contain all attributes for all products in all
store views, including translations. Feed data can be controlled via
configuration mappings of product attributes to a xml destination XPath
and a configured callback method.

Create a PIM model that, given an array of product ids, will generate a
MageMaster feed.

Create a PIM Attribute model to encapsulate data related to a specific
product attribute to be added to the feed.

Create PIM Attribute factory to build out PIM Attributes based on a
product and product attribute using any configured mapping callbacks.

Create a PIM Product model to encapsulate data related to a single item
in the feed.

Create a PIM Product collection to hold a collection of PIM Product
models. Collection will key the products by sku, client id and catalog
id. Collection also allows for a PIM Product in the collection to be
looked up by a product.

Also includes an upgrade to the dom submodule expanding upon the
capabilities of TrueAction_Dom_Document::setNode, allowing for
attributes to be included in the specified path.