Visibility Mapping, Partial Product Save Data Loss
When creating the collections of products to update, load all product
attributes for the collection instead of just the entity id. This
prevents a subtle issue in Magento where the only attribute set in the
product's origData is entity_id, causing partial data saves to wipe out
some previously set attributes, like visibility.
Clean up the mapping callback for visibility.