US7260 Customer Service Module - Epiphany Login
Adds an admin controller for customer service login from Epiphany using
a token in the URL. When the token validates, log the user in using a
configured CSR admin user.

Adds the customer service rep_id to OrderCreate messages when the logged
in admin user was logged in via a token.