DE297 Preload Product Collection in Cleaner
Load a collection of all products used by the cleaner at the beginning
of the cleaning process to prevent any additional products from needing
to be loaded while cleaning products. This collection of products
will include the products that need to be cleaned as well as any
products that may be related to the products being cleaned: mentioned
in unresolved product links or configurable product relationships.

Replace any DB lookups for related products or configurable parent or
used products by lookups within the preleaded collection.

After finishing the cleaning process, save all products in the
collection of products to ensure all changes made during the cleaning
process are saved.