US4285 Admin API Connection Test Button
Add button the the admin that will allow for the API configuration to be
tested.

Create new adminhtml controller to test the API configuration.
Controller will take either the submitted configurations or the already
saved configuration and make a test AdVal request to check the
configuration. Controller action sends a JSON response back indicating
if the test was successful and, if unsuccessful, a best guess as to why
the test failed.

Add a custom API handler to get the API call results and return the JSON
response for the controller.

Update the API model to accept an API key to be passed as a parameter
when making the API call. Necessary to allow the controller to override
the configured API key.

Add translations for the response messages. Also adds config.xml for
Eb2cCore to use thranslations in the adminhtml area.