US5631 SFTP Test Connection Button
Add button to admin to allow testing of the SFTP connection using
configured or entered details.

Refactor the API test button to match SFTP test button. Lightens up the
controller by moving most of the connection testing to a helper and
leaving only methods dealing with the request/response in the
controller. Provide more precise names for methods testing the API
connection vs the SFTP connection.