DE1476 Remove Untranslatable Xml From Pim Export
Add logistics in the eb2cproduct/pim::_getFeedAttributes to take a storeid
in order to determine if the storeid is for the default store view in order
to return all map export attributes otherwise only return transtlatable attributes.
Modified a few methods in the eb2cproduct/pim class to ensure the proper store id
is passed, and to ensure no empty 'Item' node get created.

Conflicts:

	src/app/code/community/EbayEnterprise/Eb2cProduct/Model/Pim.php