DE1479 Color Export Mappings
Add pseudo-attribute mappings for color code and description. The color
code mapping is untranslated and contains the "default" value of the
color attribute option used for the product. The description node will
be translated and contain the store view specific option label of the
color.

Update the Pim model to add the store_id to the to products as they are
being processed so the color mappings get get the correct attribute
option for the store view being processed.