DE1558 Use SVC PAN Token in OCR
Update the method used to extract a PAN from an order to be able to get
a tokenized or raw PAN from the order based on a new parameter. Use the
tokenized PAN as the account id in the OCR and the raw PAN when mapping
the SVC to a tender type.