DE1674 Remove Int Cast to Prevent Int Overflow
Int cast of the increment id has the potential to overflow if the
increment id is high enough, breaking the increments. Letting PHP handle
the type conversion itself will result in the increment being cast to an
integer when it can be done safely, or a float in all other cases. So,
to safely ensure the increment ids are generated, letting PHP handle the
conversion and then converting the incremented value back to a string
via sprintf with a format that will result in an string of integers.