DE1774 Lookup Gift Card Balance from My Account
Add the link to the Gift Cards page back into the My Account page
navigation. Replace the template with one that includes the PIN and PAN
fields. Remove the ability to "Redeem" gift cards in My Account. This is
currently done via a block override to make `canRedeem` always return
false. Should some form of this functionality be added later, this
override can simply be removed to re-enable the functionality in the
template.

Move Eb2cPayment Template to Base/Default. As the template and layout
files need to be the same for the enterprise and rwd themes, this
allows the theme fallback to get the template and layout files into both
themes instead of duplicating files in both themes.