DE1884 Remove 'null' Value From 'CharSet' Node
Rename eb2corder/create::_xsdString method to eb2corder/
create::_restrictText method and change the return signature
to return a DOMText node with the XSD valid string.
Fix broke tests. In eb2corder/create::_buildBrowserData method
make sure to passed in the $maxLength of 50 as defined in the
XSD documentation for 'CharSet' node and null as the $default
parameter.