DE2122 Add PayPal Authorization Response Code
Test first, then implemented a method in order create class
that takes a sales/order_payment instance as first parameter
and a payment method string as second parameter. For now
the method simply return the literal string 'APPROVED' however,
as we get more specification in the future the passed in sales/order_payment
and payment method parameters will use to determine the real
response code for the PayPal payment.