US2162 Consistent Request ID in Payment and OCR
Use unique ids for payment requests, SVC and PayPal and capture the
request ids used to authorize the payment methods.

In the OCR payment data, use the captured request ids to populate the
PaymentRequestId nodes for SVC, PayPal and PBridge payments.

Includes patch requred for PBridge to make the request ID avaiable in
Magento.