DE2159 Export GiftTenderCode and MaxGCAmount
Add upgrade script to make the gift_card_tender_code
attribute a required field in admin and also add max
length validation class on the attribute. Add MaxGCAmount
mapping to export MaxGCAmount as a ‘CustomAttributes’ node
when generating the item master feed. Add logistics not to
create the gift card product in the  item master feed when
both the ‘open amount max’ and the largest ‘card amount’
data are not provided on the product.

Conflicts:

	src/app/code/community/EbayEnterprise/ProductExport/etc/config.xml