DE2314 Capture Rounding Errors in PayPal Requests
When calculating a line item unit price, always round the price down to
the nearest penny. After calculating all of the line item totals, check
that the sum of the item totals matches the quote subtotal. If it
doesn't, capture the difference in a promotional rounding adjustment
line.

Use translations for the Gift Wrap line item name.

Move some tests of the abstract paypal class into the abstract tests
instead of the set express tests.

Conflicts:
	src/app/code/community/EbayEnterprise/Eb2cPayment/Test/Model/Paypal/AbstractTest.php
	src/app/locale/en_US/EbayEnterprise_Eb2cPayment.csv