DE2506 Send Amount as AmountAuthorized in OCR
The payment method is no longer 'authorized' for PayPal
causing the payment 'amount_authorized' data not to be set
in `sales/order_payment` object. This causes the
'AmountAuthorized' node value in OCR to be sent as 0.0 value.
This causes OMS to cancel the payment of the order when the
`AmountAuthorized` node is 0.00. To remedy this issue we are
sending the order grand total as the amount authorized for the
PayPal payment.