DE2381 Fix `ProductLinks` not Importing
The product import integration test uncovers a bug that is the
cause of the eb2cproduct installation script solution.
Apparently, no data type was added when creating the product
EAV attribute causing it to always create the attribute as a
`varchar ` attribute in the `catalog_product_entity_varchar`
table. As a result of this the serialize array of data being
stored for the attribute `unresolved_product_links` is now
truncated and when the cleaner try to `unserialize` the data
in the `unresolved_product_links`  attribute the data is no
longer there.