US16295 AMQP Via the ROM-SDK
Remove dependency on magento-amqp module. AMQP messages will now be
handled by the ROM-SDK.

Add a new EbayEnterprise_Amqp module to interact with the SDK for
consuming AMQP messages.

Add AMQP configuration - added to system.xml in the new amqp module and
to rom.xml.sample. Admin includes a "test connection" button that can be
used to verify a connection the AMQP server as well as a message
displaying the timestamp from the last test message received.

Remove code from the Eb2cOrder module that was interacting with AMQP or
the now defunct magento-amqp module. Eb2cOrder methods dealing directly
with handling of AMQP messages left in place to be repurposed to use the
SDK payloads.

Some refactoring of Eb2cCore API and SFTP test buttons to make it more
reusable by the EbayEnterpise_Amqp module.