US20525 Multishipping Module
Module to support multi-shipping checkout with ROM.

Create a single order for multi-shipping-address checkout.

Add fields to the order item to hold a reference to the address the item
ships to.

Add fields to the order address to hold data necessary for shipping
addresses to be used in the ORC - some totals and shipping data.

Override multishipping checkout model to use the quote service used in
onepage checkout.

Update the quote service to accommodate multiple shipping addresses. All
addresses on the quote will be converted to an order address and linked
to the order. Any items on a quote address will be linked to the
corresponding order address.

Override order model to work with multiple shipping addresses.

Override order address model to be able to manage items shipping to that
address.

Override order item model to associate it with an order address.