US35395 Additional Quantity Calculations
Add calculations for determining total quantity of an item (unique sku)
within a set of items and getting a key => value pairing of skus to
total quantities within a set of items.

Refactor calculation methods out of the inventory "Data" helper and into
a specialized helper for just doing these calculations.